component data is being passed down to components
